Prevent blacklisting bots

+++ src/commands/blacklist/blacklist-remove-sub-command.ts
@@ -40,6 +40,15 @@ export class BlacklistRemoveSubCommand {
             return;
         }
 
+        if (target.bot) {
+            let embed = new MessageEmbed()
+                .setTitle('Invalid Usage!')
+                .setDescription('You cannot blacklist a bot!')
+                .setColor(Config.colors.error);
+            await channel.send(embed);
+            return;
+        }
+
         await this.blacklistRepo.removeBlacklist(msg.guild.id, target.id);
 
         let embed = new MessageEmbed()
